Conflicts:
	cocos/2d/CCNode.cpp
	cocos/gui/UIHelper.cpp
	cocos/gui/UIInputManager.cpp
	cocos/gui/UILayout.cpp
	cocos/gui/UIListView.cpp
	cocos/gui/UIPageView.cpp
	cocos/gui/UIWidget.cpp
	cocos/gui/UIWidget.h

+++ cocos/editor-support/cocostudio/CCActionNode.cpp
@@ -399,7 +399,7 @@ int ActionNode::getLastFrameIndex()
 			continue;
 		}
 		bFindFrame = true;
-		int lastInex = cArray->count() - 1;
+		ssize_t lastInex = cArray->count() - 1;
 		ActionFrame* frame = (ActionFrame*)(cArray->getObjectAtIndex(lastInex));
 		int iFrameIndex = frame->getFrameIndex();
 
@@ -428,7 +428,7 @@ bool ActionNode::updateActionToTimeLine(float fTime)
 		{
 			continue;
 		}
-		int frameCount = cArray->count();
+		ssize_t frameCount = cArray->count();
 		for (int i = 0; i < frameCount; i++)
 		{
 			ActionFrame* frame = (ActionFrame*)(cArray->getObjectAtIndex(i));
